requirements.txt
streamlit
pandas
speechrecognition
vosk
sounddevice
pyttsx3
requests
python-dotenv
gunicorn
flask